Move `InfiniteScrollTriggerDeterminable` and implementation to a separate file. Create view modifier for measuring height.
