+++ WooCommerce/Classes/POS/Presentation/ItemListView.swift
@@ -8,7 +8,7 @@ struct ItemListView: View {
 
     @EnvironmentObject var posModel: PointOfSaleAggregateModel
 
-    @StateObject private var infiniteScrollTriggerDeterminer = InfiniteScrollTriggerDeterminer()
+    @StateObject private var infiniteScrollTriggerDeterminer = ThresholdInfiniteScrollTriggerDeterminer()
     @State private var showSimpleProductsModal: Bool = false
     private var itemListState: ItemListState {
         posModel.itemsViewState.itemsStack.root
